+++ php/php.editor/test/unit/src/org/netbeans/modules/php/editor/typinghooks/PhpTypinghooksTestBase.java
@@ -70,7 +70,7 @@ public abstract  class PhpTypinghooksTestBase extends PHPTestBase {
             String option = entry.getKey();
             Object value = entry.getValue();
             if (value instanceof Integer) {
-                prefs.putInt(option, ((Integer)value).intValue());
+                prefs.putInt(option, (Integer) value);
             }
             else if (value instanceof String) {
                 prefs.put(option, (String)value);
